# Improve getBlockEditingMode() and useAppender() performance