## What?
Fixes a performance regression introduced in https://github.com/WordPress/gutenberg/pull/51148#issuecomment-1584852322.

## How?
There's two fixes / issues at play here:

- `useAppender()` will only show a default appender in a container block when the container block is selected or at the document level when there is no selection. But `getTemplateLock()`, `getBlockEditingMode()`, and `__unstableGetEditorMode()` are called regardless of whether the appender is shown or not. This is a little wasteful. We can fix this by swapping the order of the `isParentSelected` and `isInserterHidden` checks 

- `getBlockEditingMode()` is memoised using `createSelector()` from `rememo`. I did this because `getBlockEditingMode()` will recurse through the block's ancestors. This is problematic though because [`rememo` will iterate through every cached entry](https://github.com/aduth/rememo/blob/4c4a00e023c8e761d3b354dee56885eb8bf06fd1/rememo.js#L238-L264) when the selector is called which in this case is significant as there will be one cache entry per block. The fix is to not use `createSelector()`. In most cases the depth of the block tree (number of ancestors) will be less than the number of blocks.

## Testing Instructions
Nothing specific. Check that the block appender still appears when expected at the document level and in container blocks.

## Benchmark

`trunk`

```
Typing:
Average time to type character: 38.77ms
Slowest time to type character: 108.92ms
Fastest time to type character: 33.74ms
```

This branch

```
Typing:
Average time to type character: 29.92ms
Slowest time to type character: 77.7ms
Fastest time to type character: 25.65ms
```