@youknowriad 2023-06-21
It's very cool to see that the graphs in codevitals.io are back to their previous state approximatively.